requirements.txt
python-dotenv==1.1.1
sqlalchemy==2.0.43
jinja2==3.1.2
